Organize requirement checkers into folders
