+++ requirement-generator/checkers/colleges/checkers-college-he.ts
@@ -1,4 +1,4 @@
-import { Course } from '../types';
+import { Course } from '../../types';
 
 const humanEcologyCredits = (course: Course): boolean => course.acadGroup.includes('HE');
