+++ requirement-generator/checkers/colleges/checkers-college-il.ts
@@ -1,5 +1,5 @@
-import { Course } from '../types';
-import { courseMatchesCode, courseMatchesCodeOptions, courseIsFWS } from './checkers-common';
+import { Course } from '../../types';
+import { courseMatchesCode, courseMatchesCodeOptions, courseIsFWS } from '../checkers-common';
 
 const ilrCoreRequirements = (course: Course): boolean => (
   courseMatchesCode(course, 'ILROB 1220')
