+++ requirement-generator/checkers/majors/checkers-major-aem.ts
@@ -1,5 +1,5 @@
-import { Course } from '../types';
-import { courseMatchesCode, courseMatchesCodeOptions } from './checkers-common';
+import { Course } from '../../types';
+import { courseMatchesCode, courseMatchesCodeOptions } from '../checkers-common';
 
 const aemManagementRequirements = (course: Course): boolean => (
   courseMatchesCode(course, 'AEM 2200')
