+++ requirement-generator/checkers/majors/checkers-major-govt.ts
@@ -1,5 +1,5 @@
-import { Course } from '../types';
-import { courseMatchesCode, courseMatchesCodeOptions } from './checkers-common';
+import { Course } from '../../types';
+import { courseMatchesCode, courseMatchesCodeOptions } from '../checkers-common';
 
 const governmentIntroductoryGovernmentCourses = (course: Course): boolean => courseMatchesCodeOptions(
   course,
