+++ requirement-generator/checkers/majors/checkers-major-isst.ts
@@ -1,5 +1,5 @@
-import { Course } from '../types';
-import { courseMatchesCode, courseMatchesCodeOptions } from './checkers-common';
+import { Course } from '../../types';
+import { courseMatchesCode, courseMatchesCodeOptions } from '../checkers-common';
 
 const isstProbabilityStatisticsAndOptimization = (course: Course): boolean => (
   courseMatchesCode(course, 'ORIE 3300') || courseMatchesCode(course, 'ORIE 3500')
