+++ requirement-generator/checkers/majors/checkers-major-me.ts
@@ -1,5 +1,5 @@
-import { Course } from '../types';
-import { courseMatchesCode, courseMatchesCodeOptions } from './checkers-common';
+import { Course } from '../../types';
+import { courseMatchesCode, courseMatchesCodeOptions } from '../checkers-common';
 
 const mechanicalEngineeringEngineeringDistribution = (course: Course): boolean => (
   courseMatchesCode(course, 'ENGRD 2020')
